pyproject.toml
[project]
name = "chat-interface"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-genai>=1.65.0",
    "gtts>=2.5.4",
    "mcp>=1.26.0",
    "numpy>=2.2.6",
    "python-dotenv>=1.2.1",
    "sounddevice>=0.5.5",
    "soundfile>=0.13.1",
]

[dependency-groups]
dev = [
    "ruff>=0.15.4",
    "ty>=0.0.19",
]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B", "A", "SIM", "TCH"]

[tool.ty.environment]
python-version = "3.10"
